GraphHandler class was turned into two classes: PreparationHandler and GraphManipulation
