+++ src/analysis/statistical_analysis.py
@@ -9,7 +9,7 @@ import pandas as pd
 from src import PROJECT_PATH
 from src.analysis.analysis_handler import AnalysisHandler
 from src.core.flood_wave_extractor import FloodWaveExtractor
-from src.core.graph_handler import GraphHandler
+from src.core.graph_manipulation import GraphManipulation
 from src.core.slope_calculator import SlopeCalculator
 from src.data.dataloader import Dataloader
 from src.selection.selection import Selection
@@ -37,7 +37,7 @@ class StatisticalAnalysis:
                     "end_date": f'{i}-12-31',
                     "gauge_pairs": gauge_pairs,
                     "folder_name": folder_name}
-            graph = GraphHandler.create_directed_graph(**args)
+            graph = GraphManipulation.create_directed_graph(**args)
 
             velocities = StatisticalAnalysis.calculate_all_velocities(joined_graph=graph)
             mean_velocity = np.mean(velocities)
@@ -78,7 +78,7 @@ class StatisticalAnalysis:
                 "gauge_pairs": gauge_pairs,
                 "folder_name": folder_name
             }
-            graph = GraphHandler.create_directed_graph(**args_create)
+            graph = GraphManipulation.create_directed_graph(**args_create)
 
             gauges_dct = AnalysisHandler.get_node_colors_in_given_period(gauges=gauges,
                                                                          folder_name=folder_name,
@@ -260,7 +260,7 @@ class StatisticalAnalysis:
                 "gauge_pairs": gauge_pairs,
                 "folder_name": folder_name
             }
-            graph = GraphHandler.create_directed_graph(**args_create)
+            graph = GraphManipulation.create_directed_graph(**args_create)
 
             select_all_in_interval = Selection.select_only_in_interval(joined_graph=graph,
                                                                        start_station=start_station,
@@ -366,7 +366,7 @@ class StatisticalAnalysis:
                 "end_date": end_date,
                 "gauge_pairs": gauge_pairs,
                 "folder_name": folder_name}
-        graph = GraphHandler.create_directed_graph(**args)
+        graph = GraphManipulation.create_directed_graph(**args)
 
         extractor = FloodWaveExtractor(joined_graph=graph)
         extractor.get_flood_waves()
